Be more careful when updating ccaches
Be more careful when updating credential caches.  Specifically, try to
avoid reinitializing DIR:-type ccache collections, which overwrites the
previously-set primary ccache.  This means that if we have a switchable
type, walk it as a collection, and make a note if we found anything.
Then, check if we already have a ccache for the client.  If we have such
a ccache, just use it.  If not, generate a new one, and if we didn't
have any at all before, make it the primary.