Limit what we clean out of DIR: ccaches
Limit the set of things we remove when we're cleaning up DIR: ccaches to
just the "primary" file and files whose names start with "tkt".  If
there's anything else in there that we removed before but don't know, it
probably didn't belong there, so we might as well just leave that mess.