Handle creating with non-pattern names
When called to create a ccache, accept names which don't contain a
suffix which would normally be used to generate a unique name.  When we
get one of those, if it's a file, check that it's either a file we
already own, or that we own the directory in which we'll be creating it.
Likewise, if it's a directory, check that we either own it or its
parent.  If it's a file, don't use O_EXCL when opening it.  If it's a
keyring, merely add the keyring if it doesn't exist.